site/sfguides/src/build-multi-tenant-ai-chat-application-with-cortex-rest-api/assets/requirements.txt
fastapi
uvicorn
httpx
streamlit
pydantic-settings
pyjwt
cryptography
python-dotenv
snowflake-connector-python
